refactor: move _op_function_bodies.py to functions/ directory
Consistent with LinearAttention and CausalConv ir.Function definitions
that also live in functions/. The file defines standard-ONNX ir.Function
bodies for com.microsoft custom ops (SkipLayerNorm, etc.) — exactly the
same role as the other modules in functions/.

- git mv src/mobius/_op_function_bodies.py src/mobius/functions/
- Update import in _optimizations.py (single caller)
- Export get_function + register_function_bodies from functions/__init__.py
